MailAddress: split into one-type-per-file
Extracts EmailAddressStringStrategy into its own file alongside
MailAddressStrategy.

Part of #620
