+++ src/Conjecture.Core/MailAddressStrategy.cs
@@ -7,19 +7,6 @@ using Conjecture.Core.Internal;
 
 namespace Conjecture.Core;
 
-internal sealed class EmailAddressStringStrategy : Strategy<string>
-{
-    private readonly IdentifierStrategy localPart = new(1, 6, 1, 4);
-    private readonly HostStrategy hostPart = new(2, 2);
-
-    internal override string Generate(ConjectureData data)
-    {
-        string local = localPart.Generate(data);
-        string host = hostPart.Generate(data);
-        return local + "@" + host;
-    }
-}
-
 internal sealed class MailAddressStrategy : Strategy<MailAddress>
 {
     private readonly EmailAddressStringStrategy stringStrategy = new();
